+++ R/theme_gtsummary_msk.R
@@ -32,11 +32,13 @@ theme_gtsummary_msk <- function(name = c("hot", "karissa", "ally", "mauguen",
   lst_theme_hot <- list(
     "pkgwide-str:theme_name" = "H.O.T.",
     # display a greek beta as header in tbl_regression
-    "tbl_regression-str:coef_header" = rlang::expr(ifelse(exponentiate == TRUE, "exp(\U03B2)", "\U03B2")),
-    "as_flex_table-lst:addl_cmds" = list(autofit = list(rlang::expr(flextable::font(fontname = "Calibri", part = "all"))))
+    "tbl_regression-str:coef_header" = rlang::expr(ifelse(exponentiate == TRUE, "exp(\U03B2)", "\U03B2"))
   ) %>%
     # adding compact theme (removing name, however)
     c(gtsummary::theme_gtsummary_compact(set_theme = FALSE, font_size = font_size)[-1])
+  lst_theme_hot$`as_flex_table-lst:addl_cmds`$valign <-
+    c(lst_theme_hot$`as_flex_table-lst:addl_cmds`$valign,
+      list(rlang::expr(flextable::font(fontname = "Calibri", part = "all"))))
 
   # selecting theme list
   switch(