+++ NEWS.md
@@ -1,3 +1,5 @@
+# bstfun (development version)
+
 # bstfun 0.4.9
 
 * Added gtreg, emmeans, ggsurvfit, and tidycmprsk to the template references document. (#112)